requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-generativeai
python-dotenv
langchain
textract